+++ src/constants/SlideTheme.ts
@@ -1,0 +1,6 @@
+export enum SlideTheme {
+  default = "default",
+  minimal = "minimal",
+  monomi = "monomi",
+  unnamed = "unnamed",
+}