+++ src/constants/index.ts
@@ -3,5 +3,6 @@ export * from "./Config";
 export * from "./ContextKeys";
 export * from "./General";
 export * from "./SlideLayout";
+export * from "./SlideTheme";
 export * from "./StateKeys";
 export * from "./WebViewMessages";